Move LoadBalancerServiceException to config-provisioning module
