+++ config-provisioning/src/main/java/com/yahoo/config/provision/exception/LoadBalancerServiceException.java
@@ -1,10 +1,8 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.hosted.provision.lb;
-
-import com.yahoo.config.provision.exception.TransientException;
+package com.yahoo.config.provision.exception;
 
 /**
- * Transient exception thrown on behalf of a {@link LoadBalancerService}.
+ * Transient exception thrown on behalf of a load balancer service
  *
  * @author mpolden
  */
