+++ config-provisioning/src/main/java/com/yahoo/config/provision/exception/ParentHostUnavailableException.java
@@ -1,8 +1,6 @@
 // Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.config.provision.exception;
 
-import com.yahoo.config.provision.exception.TransientException;
-
 /**
  *
  * Exception thrown when trying to activate a node that runs on a host that is not
